requirements.txt
Flask
SQLAlchemy
psycopg2
Flask-SQLAlchemy
Werkzeug
PyJWT